# Conflicts:
#	Content/SY/Enemy/Enemy03/BP_Enemy03.uasset
#	Source/PJ_MZ/Character/Enemy/EnemySpawnManager.cpp

+++ Source/PJ_MZ/MZ_Datas.h
@@ -102,7 +102,7 @@ struct FLeaderboardEntry
 	FString PlayerName;
 
 	UPROPERTY(BlueprintReadOnly)
-	int32 Score;
+	float Score;
 
 	UPROPERTY(BlueprintReadOnly)
 	float ClearTime;
